skills/youtube-to-slides/pyproject.toml
[project]
name = "yt-slides"
version = "0.1.0"
description = "Convert YouTube videos into infographic slides using Gemini AI"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.0.0",
    "youtube-transcript-api>=1.0.0",
    "google-api-python-client>=2.0.0",
    "typer>=0.15.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
yt-slides = "yt_slides.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
